Integrated SRP client API with the existing mDNS API.
The SRP and mDNS protocols should have some common layer
and API. For now it seems to be reasonable to make SRP client
to behave as mDNS implementation for Thread devices, so some
integration is necessary in order to reuse existing mDNS code.

* Replaced chip_mdns_advertiser and chip_enable_mdns arguments,
with one - chip_mdns acting as both of them.
* Added implementation of SetupHostname() method for Thread devices.
* Added MdnsImpl for the nrfconnect platform.
* Expanded AddSrpService method by possibility to pass text entries.
* Moved TextEntry struct from Mdns.h to the inet/DNS.h, as it was
declared in the src/lib/mdns and because of that inaccessible in
ThreadStackManager.
* Added ifdef to not call AdvertiseCommissioning for Thread devices,
as they are not able to do that in unprovisioned state.
* Added ThreadStackManager method for getting device EUI64.