[nrfconnect] Aligned BLE advertising to be spec compatible.
According to spec CHIP device should advertise using BLE with the
fast interval (20 ms - 60 ms) for 30 seconds and then change
interval to the slow one (150 ms - 1200 ms). Currently it is not
implemented.

* Introduced new defines in CHIPDeviceConfig.h storing minimum
and maximum intervals for slow and fast mode.
* Replaced unused SetFastAdvertisingEnabled method with SetAdvertisingMode,
as the name was confusing (method needed calling SetAdvertisingEnabled too
in order to work correctly).
* Set fast advertising mode as a default to be compatible with the spec.
* Changed for all platforms advertising intervals min and max to not
be equal.
* For nrfconnect platform added timer event changing BLE advertising
interval after 30 seconds.
* Reverted #... to restore general advertising mode.