[ble] Fixed controller not being able to commission device after first attempt failure
Sometimes it happens that if in specific moment of device
commissioning over Bluetooth LE the connection will be closed
(e.g. by turning off HCI adapter to simulate hardware problems)
the controller is not able to commission the device on the second
attempt due to incorrect state.

* Added setting BLE state to kInitialized if BLE connection
establishment failed.
* Added cleaning rendezvous data if some PASE establisment
error appeared.