[nrfconnect] Refactored configuration for the nrfconnect examples
nrfconnect platform uses Kconfig overlays and defaults in a way
that is not consistent with the nRF Connect SDK recommendations

* Created Kconfig.defaults that set default configs common
for all examples (replaced sample-defaults.conf)
* Created Kconfig.features that allows to configure set
of options within single feature using single option
* Replaced approach of adding overlays in example cmake file
with the pattern of placing prj.conf files in a
configuration/build-target/ location that allows adding config
files to the build automatically by the toolchain.
* Created debug and release configuration variants for several
examples.